Minor fixes

+++ src/riscv/linux/riscv-isa.c
@@ -6,7 +6,7 @@
 
 void cpuinfo_riscv_linux_decode_isa_from_hwcap(
 		struct cpuinfo_riscv_isa isa[restrict static 1]) {
-	unsigned long hwcap = 0;
+	uint32_t hwcap = 0;
 	cpuinfo_riscv_linux_hwcap_from_getauxval(&hwcap);
 
 	if (hwcap & CPUINFO_RISCV_LINUX_FEATURE_I) {
